Add test for failing fips lookup
Sometimes, running the irwin polling worker would cause wildfire to
crash with a failing fips lookup. The reason for this was that fips was
nil - actually, all attributes on the wildfire record were nil. This was
during the after_create callback to send texts to all nearby users, so
this was an issue. The cause turned out to be the use of
find_or_create_by: I intended for the record to be either found or
initialized, but by using create_by I triggered Wildfire's after_create
callbacks before the record had any data. I fixed this by using
find_or_initialize_by, calls new if the record is not found.