Delete user data if text comes in
This commit adds a webhook that the Twilio API calls upon receipt of a
text message from a user. If the user wants to delete their record for a
zip or all their records, they can text "DELETE ALL" or "DELETE {zip}".
The endpoint renders a TwiML response that Twilio will then pass back to
the user.

This also adds tests to verify the functionality. One task that still
needs to be completed is to validate that the webhook request is coming
from Twilio, as right now the webhook can be called from anywhere.