Update text message format sent for fires
Texts about fires currently only say the coordinates of where a fire is
reported. This isn't super helpful in a text message as you can't copy
the coordinates and it doesn't tell you much about how close the fire is
to a place you care about. This commit updates Wildfire's to_s method to
accept a zip, and if it's present then it returns the distance to that
zip instead. This makes the format of messages much more readable, going
from
KNOLL 2 (0.0 acres, 0% contained) reported at 37.16066, -122.1345
to
KNOLL 2 (0.0 acres, 0% contained), 11.88 miles from 95014