Switch from hash to hexdigest for tracking texts
To make sure I don't send duplicate texts, I used to hash each message
and compare its value across all text alerts stored in the database. I
noticed a bug once the project was deployed where duplicate text
messages would be sent despite this logic. After some digging, it
appears that the String hash function Ruby provides is not
deterministic. From the doc for Object#hash:

The hash value for an object may not be identical across invocations or
implementations of Ruby. If you need a stable identifier across Ruby
invocations and implementations you will need to generate one with a
custom method.

I'm getting around this with the Digest module to store and compare a
message's SHA1 hexdigest.